+++ src/components/Navbar.tsx
@@ -99,6 +99,7 @@ const Navbar = () => {
     });
 
     return () => observer.disconnect();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (